+++ src/strongbox/core.clj
@@ -220,6 +220,8 @@
 
 ;; addon dirs
 
+(def default-game-track-strictness true) ;; strict
+
 (defn-spec selected-addon-dir (s/or :ok ::sp/addon-dir, :no-selection nil?)
   "returns the currently selected addon directory or nil if no directories exist to select from"
   ([]
@@ -234,11 +236,14 @@
    (->> addon-dir-list (map :addon-dir) (some #{addon-dir}) nil? not)))
 
 (defn-spec add-addon-dir! nil?
-  [addon-dir ::sp/addon-dir, game-track ::sp/game-track]
-  (let [stub {:addon-dir addon-dir :game-track game-track}]
-    (when-not (addon-dir-exists? addon-dir)
-      (swap! state update-in [:cfg :addon-dir-list] conj stub))
-    nil))
+  "creates and adds an addon directory entry in the user's `:addon-dir-list`, if it doesn't already exist."
+  ([addon-dir ::sp/addon-dir, game-track ::sp/game-track]
+   (add-addon-dir! addon-dir game-track default-game-track-strictness))
+  ([addon-dir ::sp/addon-dir, game-track ::sp/game-track, strict? ::sp/strict?]
+   (let [stub {:addon-dir addon-dir :game-track game-track :strict? strict?}]
+     (when-not (addon-dir-exists? addon-dir)
+       (swap! state update-in [:cfg :addon-dir-list] conj stub)))
+   nil))
 
 ;; see also: strongbox.ui.cli/set-addon-dir! 
 (defn-spec set-addon-dir! nil?
@@ -248,7 +253,7 @@
         ;; if '_classic_' is in given path, use the classic game track
         default-game-track (if (clojure.string/index-of addon-dir "_classic_") :classic :retail)]
     (dosync ;; necessary? makes me feel better
-     (add-addon-dir! addon-dir default-game-track)
+     (add-addon-dir! addon-dir default-game-track default-game-track-strictness)
      ;; todo: this is UI logic ... consider moving to ui.cli
      (swap! state assoc-in [:cfg :selected-addon-dir] addon-dir)
      (swap! state assoc :tab-list []))) ;; todo: consider adding a watch for this as well
@@ -277,9 +282,10 @@
   ([]
    (when-let [addon-dir (selected-addon-dir)]
      (addon-dir-map addon-dir)))
-  ([addon-dir ::sp/addon-dir]
+  ([addon-dir (s/nilable ::sp/addon-dir)]
    (let [addon-dir-list (get-state :cfg :addon-dir-list)]
-     (when-not (empty? addon-dir-list)
+     (when (and addon-dir
+                (not (empty? addon-dir-list)))
        (first (filter #(= addon-dir (:addon-dir %)) addon-dir-list))))))
 
 (defn-spec set-game-track! nil?
@@ -305,14 +311,26 @@
    (when addon-dir
      (-> addon-dir addon-dir-map :game-track))))
 
-(defn-spec get-lenient-game-track ::sp/lenient-game-track
-  "returns the lenient/compound version of the currently selected game track. 
-  if `:retail` then `:retail-classic`, etc"
+(defn-spec get-game-track-strictness (s/or :addon-dir ::sp/strict?, :no-addon-dir nil?)
   []
-  (case (get-game-track)
-    :classic-retail :classic-retail
-    :classic :classic-retail
-    :retail-classic))
+  (when-let [addon-dir-map (addon-dir-map)]
+    (get addon-dir-map :strict? default-game-track-strictness)))
+
+;; todo: this is almost identical with `get-game-track`.
+;; come up with a better solution if we repeat ourselves again.
+(defn-spec set-game-track-strictness! nil?
+  "fetches the strictness level for the given `addon-dir` or the currently selected addon directory if not given."
+  ([new-strictness-level ::sp/strict?]
+   (when-let [addon-dir (selected-addon-dir)]
+     (set-game-track-strictness! new-strictness-level addon-dir)))
+  ([new-strictness-level ::sp/strict?, addon-dir ::sp/addon-dir]
+   (let [tform (fn [addon-dir-map]
+                 (if (= addon-dir (:addon-dir addon-dir-map))
+                   (assoc addon-dir-map :strict? new-strictness-level)
+                   addon-dir-map))
+         new-addon-dir-map-list (mapv tform (get-state :cfg :addon-dir-list))]
+     (swap! state update-in [:cfg] assoc :addon-dir-list new-addon-dir-map-list))
+   nil))
 
 
 ;; stateful logging
@@ -747,8 +765,9 @@
   [addon-summary]
   (binding [http/*cache* (cache)]
     (let [game-track (get-game-track) ;; scope creep, but it fits so nicely
+          strict? (get-game-track-strictness)
           wrapper (affects-addon-wrapper catalogue/expand-summary)]
-      (wrapper addon-summary game-track))))
+      (wrapper addon-summary game-track strict?))))
 
 (defn-spec check-for-update :addon/toc
   "Returns given `addon` with source updates, if any, and sets an `update?` property if a different version is available.
@@ -983,12 +1002,13 @@
         ;; when no game-track is present in the export record, use the more lenient
         ;; version of the currently selected game track.
         ;; it's better to have an addon installed with the incorrect game track then missing addons.
-        default-game-track (get-lenient-game-track)]
+        default-game-track (get-game-track)
+        strict? false]
 
     (binding [http/*cache* (cache)]
       (doseq [addon matching-addon-list
               :let [game-track (get addon :game-track default-game-track)]]
-        (when-let [expanded-addon (catalogue/expand-summary addon game-track)]
+        (when-let [expanded-addon (catalogue/expand-summary addon game-track strict?)]
           (install-addon expanded-addon))))))
 
 (defn-spec import-exported-file nil?
@@ -1077,7 +1097,7 @@
   (binding [http/*cache* (cache)]
     (if-let* [addon-summary (catalogue/parse-user-string addon-url)
               ;; game track doesn't matter when adding it to the user catalogue. prefer retail though.
-              addon (catalogue/expand-summary addon-summary :retail-classic)
+              addon (catalogue/expand-summary addon-summary :retail false)
               test-only? true
               _ (install-addon-guard addon (selected-addon-dir) test-only?)]