+++ test/strongbox/config_test.clj
@@ -122,7 +122,8 @@
                                        :game-track :classic}]}
           expected (assoc config/default-cfg
                           :addon-dir-list [{:addon-dir (str fs/*cwd*)
-                                            :game-track :classic}]
+                                            :game-track :classic
+                                            :strict? true}]
                           :selected-addon-dir (str fs/*cwd*))]
       (is (= expected (config/merge-config file-opts cli-opts))))))
 
@@ -160,6 +161,17 @@
                     :selected-addon-dir nil}]
       (is (= expected (config/handle-selected-addon-dir given))))))
 
+(deftest convert-compound-game-track
+  (let [cases [[:retail :retail]
+               [:classic :classic]
+               [:retail-classic :retail]
+               [:classic-retail :classic]
+               [:classic-tbc :classic-tbc]]]
+    (doseq [[given expected] cases]
+      (is (= expected (config/convert-compound-game-track given))))))
+
+;; ---
+
 (deftest load-settings-0.9
   (testing "a standard config file circa 0.9 is loaded and parsed as expected"
     (let [cli-opts {}
@@ -170,8 +182,8 @@
                           :selected-catalogue :short ;; new in 0.10
                           ;; :debug? true ;; removed in 0.12
                           ;; new in 0.12
-                          :addon-dir-list [{:addon-dir "/tmp/.strongbox-bar", :game-track :retail}
-                                           {:addon-dir "/tmp/.strongbox-foo", :game-track :classic}]
+                          :addon-dir-list [{:addon-dir "/tmp/.strongbox-bar", :game-track :retail, :strict? true}
+                                           {:addon-dir "/tmp/.strongbox-foo", :game-track :classic, :strict? true}]
                           ;; new in 1.0
                           :catalogue-location-list (:catalogue-location-list config/default-cfg)
 
@@ -199,8 +211,8 @@
           expected {:cfg {:gui-theme :light ;; new in 0.11
                           :selected-catalogue :full ;; new in 0.10
                           ;; :debug? true ;; removed in 0.12
-                          :addon-dir-list [{:addon-dir "/tmp/.strongbox-bar", :game-track :retail}
-                                           {:addon-dir "/tmp/.strongbox-foo", :game-track :classic}]
+                          :addon-dir-list [{:addon-dir "/tmp/.strongbox-bar", :game-track :retail, :strict? true}
+                                           {:addon-dir "/tmp/.strongbox-foo", :game-track :classic, :strict? true}]
                           ;; new in 1.0
                           :catalogue-location-list (:catalogue-location-list config/default-cfg)
 
@@ -229,8 +241,8 @@
           expected {:cfg {:gui-theme :dark ;; new in 0.11
                           :selected-catalogue :full ;; new in 0.10
                           ;;:debug? true ;; removed in 0.12
-                          :addon-dir-list [{:addon-dir "/tmp/.strongbox-bar", :game-track :retail}
-                                           {:addon-dir "/tmp/.strongbox-foo", :game-track :classic}]
+                          :addon-dir-list [{:addon-dir "/tmp/.strongbox-bar", :game-track :retail, :strict? true}
+                                           {:addon-dir "/tmp/.strongbox-foo", :game-track :classic, :strict? true}]
                           ;; new in 1.0
                           :catalogue-location-list (:catalogue-location-list config/default-cfg)
 
@@ -245,6 +257,7 @@
                     :file-opts {:gui-theme :dark
                                 :selected-catalogue :full
                                 :debug? true
+                                ;; todo: shouldn't these have `:strict?` in them?
                                 :addon-dir-list [{:addon-dir "/tmp/.strongbox-bar", :game-track :retail}
                                                  {:addon-dir "/tmp/.strongbox-foo", :game-track :classic}]}
                     :etag-db {}}]
@@ -259,8 +272,8 @@
           expected {:cfg {:gui-theme :dark ;; new in 0.11
                           :selected-catalogue :full ;; new in 0.10
                           ;;:debug? true ;; removed in 0.12
-                          :addon-dir-list [{:addon-dir "/tmp/.strongbox-bar", :game-track :retail}
-                                           {:addon-dir "/tmp/.strongbox-foo", :game-track :classic}]
+                          :addon-dir-list [{:addon-dir "/tmp/.strongbox-bar", :game-track :retail, :strict? true}
+                                           {:addon-dir "/tmp/.strongbox-foo", :game-track :classic, :strict? true}]
                           ;; new in 1.0
                           :catalogue-location-list (:catalogue-location-list config/default-cfg)
 
@@ -288,8 +301,8 @@
           expected {:cfg {:gui-theme :dark ;; new in 0.11
                           :selected-catalogue :full ;; new in 0.10
                           ;;:debug? true ;; removed in 0.12
-                          :addon-dir-list [{:addon-dir "/tmp/.strongbox-bar", :game-track :retail}
-                                           {:addon-dir "/tmp/.strongbox-foo", :game-track :classic}]
+                          :addon-dir-list [{:addon-dir "/tmp/.strongbox-bar", :game-track :retail, :strict? true}
+                                           {:addon-dir "/tmp/.strongbox-foo", :game-track :classic, :strict? true}]
 
                           ;; new in 1.0
                           :catalogue-location-list (:catalogue-location-list config/default-cfg)
@@ -323,8 +336,9 @@
           expected {:cfg {:gui-theme :dark ;; new in 0.11
                           :selected-catalogue :full ;; new in 0.10
                           ;;:debug? true ;; removed in 0.12
-                          :addon-dir-list [{:addon-dir "/tmp/.strongbox-bar", :game-track :retail-classic} ;; compound game tracks added in 3.1
-                                           {:addon-dir "/tmp/.strongbox-foo", :game-track :classic}]
+                          :addon-dir-list [;;{:addon-dir "/tmp/.strongbox-bar", :game-track :retail-classic} ;; compound game tracks added in 3.1
+                                           {:addon-dir "/tmp/.strongbox-bar", :game-track :retail :strict? false} ;; compound game tracks removed in 4.1
+                                           {:addon-dir "/tmp/.strongbox-foo", :game-track :classic :strict? true}]
 
                           ;; new in 1.0
                           :catalogue-location-list (:catalogue-location-list config/default-cfg)
@@ -360,8 +374,9 @@
           expected {:cfg {:gui-theme :dark-green ;; new in 0.11, `:dark-green` new in 3.2.0
                           :selected-catalogue :full ;; new in 0.10
                           ;;:debug? true ;; removed in 0.12
-                          :addon-dir-list [{:addon-dir "/tmp/.strongbox-bar", :game-track :retail-classic} ;; compound game tracks added in 3.1
-                                           {:addon-dir "/tmp/.strongbox-foo", :game-track :classic}]
+                          :addon-dir-list [;;{:addon-dir "/tmp/.strongbox-bar", :game-track :retail-classic} ;; compound game tracks added in 3.1
+                                           {:addon-dir "/tmp/.strongbox-bar", :game-track :retail :strict? false} ;; compound game tracks removed in 4.1
+                                           {:addon-dir "/tmp/.strongbox-foo", :game-track :classic :strict? true}]
 
                           ;; new in 1.0
                           :catalogue-location-list (:catalogue-location-list config/default-cfg)
@@ -378,6 +393,43 @@
                                 :selected-catalogue :full
                                 :addon-dir-list [{:addon-dir "/tmp/.strongbox-bar", :game-track :retail-classic}
                                                  {:addon-dir "/tmp/.strongbox-foo", :game-track :classic}]
+                                :selected-addon-dir "/tmp/.strongbox-foo"
+
+                                ;; new in 1.0
+                                :catalogue-location-list (:catalogue-location-list config/default-cfg)
+
+                                :preferences {:addon-zips-to-keep 3}}
+
+                    :etag-db {}}]
+      (is (= expected (config/load-settings cli-opts cfg-file etag-db-file))))))
+
+(deftest load-settings-4.1
+  (testing "a standard config file circa 4.1 is loaded and parsed as expected"
+    (let [cli-opts {}
+          cfg-file (fixture-path "user-config-4.1.json")
+          etag-db-file (fixture-path "empty-map.json")
+
+          expected {:cfg {:gui-theme :dark-green ;; new in 0.11, `:dark-green` new in 3.2.0
+                          :selected-catalogue :full ;; new in 0.10
+                          ;;:debug? true ;; removed in 0.12
+                          :addon-dir-list [{:addon-dir "/tmp/.strongbox-bar", :game-track :classic-tbc :strict? true} ;; `:classic-tbc` and `:strict?` added in 4.1
+                                           {:addon-dir "/tmp/.strongbox-foo", :game-track :retail :strict? false}]
+
+                          ;; new in 1.0
+                          :catalogue-location-list (:catalogue-location-list config/default-cfg)
+
+                          ;; new in 0.12
+                          ;; moved to :cfg in 1.0
+                          :selected-addon-dir "/tmp/.strongbox-foo"
+
+                          ;; new in 3.1.0
+                          :preferences {:addon-zips-to-keep 3}}
+
+                    :cli-opts {}
+                    :file-opts {:gui-theme :dark-green
+                                :selected-catalogue :full
+                                :addon-dir-list [{:addon-dir "/tmp/.strongbox-bar", :game-track :classic-tbc, :strict? true}
+                                                 {:addon-dir "/tmp/.strongbox-foo", :game-track :retail, :strict? false}]
                                 :selected-addon-dir "/tmp/.strongbox-foo"
 
                                 ;; new in 1.0